refactored non-GUI related tests
